Remove in memory DAOs and configuration option to disable slick
