+++ README.md
@@ -8,10 +8,11 @@ It uses the [play-slick](https://github.com/playframework/play-slick) library fo
 
 ## Example
 
-Currently, there is no live example of this template.
 [![Deploy to Heroku](https://www.herokucdn.com/deploy/button.png)](https://heroku.com/deploy)
 
 (The "Build App" phase will take a few minutes)
+
+Currently, there is no live example of this template.
 
 ## Features
 
@@ -33,14 +34,6 @@ Consulate the [Silhouette documentation](http://silhouette.mohiva.com/docs) for 
 The template stores all authentication information in a database via [Slick](http://slick.typesafe.com/) It uses an in memory [H2](www.h2database.com/) database by default.
 
 In order to use another database supported by Slick, you need to change the driver in your application.conf and add the corresponding JDBC driver to your dependencies. The [Play Slick documentation](https://www.playframework.com/documentation/2.4.x/PlaySlick) has more information about database configuration.
-
-To disable database access and fall back to in memory data storage via hash maps, edit the play configuration:
-
-    silhouette.seed.db.useSlick=false
-
-### Documentation
-
-Consulate the [Silhouette documentation](http://silhouette.mohiva.com/docs) for more information. If you need help with the integration of Silhouette into your project, don't hesitate and ask questions in our [mailing list](https://groups.google.com/forum/#!forum/play-silhouette) or on [Stack Overflow](http://stackoverflow.com/questions/tagged/playframework).
 
 ## Activator
