+++ app/models/daos/OAuth1InfoDAO.scala
@@ -3,18 +3,36 @@ package models.daos
 import com.mohiva.play.silhouette.api.LoginInfo
 import com.mohiva.play.silhouette.impl.daos.DelegableAuthInfoDAO
 import com.mohiva.play.silhouette.impl.providers.OAuth1Info
-import models.daos.OAuth1InfoDAO._
 import play.api.libs.concurrent.Execution.Implicits._
 
-import scala.collection.mutable
 import scala.concurrent.Future
 
 /**
  * The DAO to store the OAuth1 information.
- *
- * Note: Not thread safe, demo only.
  */
-class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] {
+class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] with DAOSlick {
+
+  import driver.api._
+
+  protected def oAuth1InfoQuery(loginInfo: LoginInfo) = for {
+    dbLoginInfo <- loginInfoQuery(loginInfo)
+    dbOAuth1Info <- slickOAuth1Infos if dbOAuth1Info.loginInfoId === dbLoginInfo.id
+  } yield dbOAuth1Info
+  
+  // Use subquery workaround instead of join to get authinfo because slick only supports selecting
+  // from a single table for update/delete queries (https://github.com/slick/slick/issues/684).
+  protected def oAuth1InfoSubQuery(loginInfo: LoginInfo) =
+    slickOAuth1Infos.filter(_.loginInfoId in loginInfoQuery(loginInfo).map(_.id))
+
+  protected def addAction(loginInfo: LoginInfo, authInfo: OAuth1Info) =
+    loginInfoQuery(loginInfo).result.head.flatMap { dbLoginInfo =>
+      slickOAuth1Infos += DBOAuth1Info(None, authInfo.token, authInfo.secret, dbLoginInfo.id.get)
+    }.transactionally
+
+  protected def updateAction(loginInfo: LoginInfo, authInfo: OAuth1Info) =
+    oAuth1InfoSubQuery(loginInfo).
+      map(dbOAuthInfo => (dbOAuthInfo.token, dbOAuthInfo.secret)).
+      update((authInfo.token, authInfo.secret))
 
   /**
    * Finds the auth info which is linked with the specified login info.
@@ -23,7 +41,10 @@ class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] {
    * @return The retrieved auth info or None if no auth info could be retrieved for the given login info.
    */
   def find(loginInfo: LoginInfo): Future[Option[OAuth1Info]] = {
-    Future.successful(data.get(loginInfo))
+    val result = db.run(oAuth1InfoQuery(loginInfo).result.headOption)
+    result.map { dbOAuth1InfoOption =>
+      dbOAuth1InfoOption.map(dbOAuth1Info => OAuth1Info(dbOAuth1Info.token, dbOAuth1Info.secret))
+    }
   }
 
   /**
@@ -33,10 +54,8 @@ class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] {
    * @param authInfo The auth info to add.
    * @return The added auth info.
    */
-  def add(loginInfo: LoginInfo, authInfo: OAuth1Info): Future[OAuth1Info] = {
-    data += (loginInfo -> authInfo)
-    Future.successful(authInfo)
-  }
+  def add(loginInfo: LoginInfo, authInfo: OAuth1Info): Future[OAuth1Info] = 
+    db.run(addAction(loginInfo, authInfo)).map(_ => authInfo)
 
   /**
    * Updates the auth info for the given login info.
@@ -45,10 +64,8 @@ class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] {
    * @param authInfo The auth info to update.
    * @return The updated auth info.
    */
-  def update(loginInfo: LoginInfo, authInfo: OAuth1Info): Future[OAuth1Info] = {
-    data += (loginInfo -> authInfo)
-    Future.successful(authInfo)
-  }
+  def update(loginInfo: LoginInfo, authInfo: OAuth1Info): Future[OAuth1Info] = 
+    db.run(updateAction(loginInfo, authInfo)).map(_ => authInfo)
 
   /**
    * Saves the auth info for the given login info.
@@ -61,10 +78,12 @@ class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] {
    * @return The saved auth info.
    */
   def save(loginInfo: LoginInfo, authInfo: OAuth1Info): Future[OAuth1Info] = {
-    find(loginInfo).flatMap {
-      case Some(_) => update(loginInfo, authInfo)
-      case None => add(loginInfo, authInfo)
-    }
+    val query = loginInfoQuery(loginInfo).joinLeft(slickOAuth1Infos).on(_.id === _.loginInfoId)
+    val action = query.result.head.flatMap {
+      case (dbLoginInfo, Some(dbOAuth1Info)) => updateAction(loginInfo, authInfo)
+      case (dbLoginInfo, None)               => addAction(loginInfo, authInfo)
+    }.transactionally
+    db.run(action).map(_ => authInfo)
   }
 
   /**
@@ -73,19 +92,6 @@ class OAuth1InfoDAO extends DelegableAuthInfoDAO[OAuth1Info] {
    * @param loginInfo The login info for which the auth info should be removed.
    * @return A future to wait for the process to be completed.
    */
-  def remove(loginInfo: LoginInfo): Future[Unit] = {
-    data -= loginInfo
-    Future.successful(())
-  }
-}
-
-/**
- * The companion object.
- */
-object OAuth1InfoDAO {
-
-  /**
-   * The data store for the OAuth1 info.
-   */
-  var data: mutable.HashMap[LoginInfo, OAuth1Info] = mutable.HashMap()
+  def remove(loginInfo: LoginInfo): Future[Unit] =
+    db.run(oAuth1InfoSubQuery(loginInfo).delete).map(_ => ())
 }
