+++ app/models/daos/PasswordInfoDAO.scala
@@ -3,17 +3,38 @@ package models.daos
 import com.mohiva.play.silhouette.api.LoginInfo
 import com.mohiva.play.silhouette.api.util.PasswordInfo
 import com.mohiva.play.silhouette.impl.daos.DelegableAuthInfoDAO
-import models.daos.PasswordInfoDAO._
 import play.api.libs.concurrent.Execution.Implicits._
 
-import scala.collection.mutable
 import scala.concurrent.Future
 
 /**
  * The DAO to store the password information.
  */
-class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
+class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] with DAOSlick{
 
+  import driver.api._
+
+  protected def passwordInfoQuery(loginInfo: LoginInfo) = for {
+    dbLoginInfo <- loginInfoQuery(loginInfo)
+    dbPasswordInfo <- slickPasswordInfos if dbPasswordInfo.loginInfoId === dbLoginInfo.id
+  } yield dbPasswordInfo
+  
+  // Use subquery workaround instead of join to get authinfo because slick only supports selecting
+  // from a single table for update/delete queries (https://github.com/slick/slick/issues/684).
+  protected def passwordInfoSubQuery(loginInfo: LoginInfo) =
+    slickPasswordInfos.filter(_.loginInfoId in loginInfoQuery(loginInfo).map(_.id))
+
+  protected def addAction(loginInfo: LoginInfo, authInfo: PasswordInfo) =
+    loginInfoQuery(loginInfo).result.head.flatMap { dbLoginInfo =>
+      slickPasswordInfos +=
+        DBPasswordInfo(authInfo.hasher, authInfo.password, authInfo.salt, dbLoginInfo.id.get)
+    }.transactionally
+    
+  protected def updateAction(loginInfo: LoginInfo, authInfo: PasswordInfo) =
+    passwordInfoSubQuery(loginInfo).
+      map(dbPasswordInfo => (dbPasswordInfo.hasher, dbPasswordInfo.password, dbPasswordInfo.salt)).
+      update((authInfo.hasher, authInfo.password, authInfo.salt))
+  
   /**
    * Finds the auth info which is linked with the specified login info.
    *
@@ -21,7 +42,10 @@ class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
    * @return The retrieved auth info or None if no auth info could be retrieved for the given login info.
    */
   def find(loginInfo: LoginInfo): Future[Option[PasswordInfo]] = {
-    Future.successful(data.get(loginInfo))
+    db.run(passwordInfoQuery(loginInfo).result.headOption).map { dbPasswordInfoOption =>
+      dbPasswordInfoOption.map(dbPasswordInfo => 
+        PasswordInfo(dbPasswordInfo.hasher, dbPasswordInfo.password, dbPasswordInfo.salt))
+    }
   }
 
   /**
@@ -31,10 +55,8 @@ class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
    * @param authInfo The auth info to add.
    * @return The added auth info.
    */
-  def add(loginInfo: LoginInfo, authInfo: PasswordInfo): Future[PasswordInfo] = {
-    data += (loginInfo -> authInfo)
-    Future.successful(authInfo)
-  }
+  def add(loginInfo: LoginInfo, authInfo: PasswordInfo): Future[PasswordInfo] =
+    db.run(addAction(loginInfo, authInfo)).map(_ => authInfo)
 
   /**
    * Updates the auth info for the given login info.
@@ -43,10 +65,8 @@ class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
    * @param authInfo The auth info to update.
    * @return The updated auth info.
    */
-  def update(loginInfo: LoginInfo, authInfo: PasswordInfo): Future[PasswordInfo] = {
-    data += (loginInfo -> authInfo)
-    Future.successful(authInfo)
-  }
+  def update(loginInfo: LoginInfo, authInfo: PasswordInfo): Future[PasswordInfo] = 
+    db.run(updateAction(loginInfo, authInfo)).map(_ => authInfo)
 
   /**
    * Saves the auth info for the given login info.
@@ -59,10 +79,12 @@ class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
    * @return The saved auth info.
    */
   def save(loginInfo: LoginInfo, authInfo: PasswordInfo): Future[PasswordInfo] = {
-    find(loginInfo).flatMap {
-      case Some(_) => update(loginInfo, authInfo)
-      case None => add(loginInfo, authInfo)
+    val query = loginInfoQuery(loginInfo).joinLeft(slickPasswordInfos).on(_.id === _.loginInfoId)
+    val action = query.result.head.flatMap {
+      case (dbLoginInfo, Some(dbPasswordInfo)) => updateAction(loginInfo, authInfo)
+      case (dbLoginInfo, None) => addAction(loginInfo, authInfo)
     }
+    db.run(action).map(_ => authInfo)
   }
 
   /**
@@ -71,19 +93,6 @@ class PasswordInfoDAO extends DelegableAuthInfoDAO[PasswordInfo] {
    * @param loginInfo The login info for which the auth info should be removed.
    * @return A future to wait for the process to be completed.
    */
-  def remove(loginInfo: LoginInfo): Future[Unit] = {
-    data -= loginInfo
-    Future.successful(())
-  }
-}
-
-/**
- * The companion object.
- */
-object PasswordInfoDAO {
-
-  /**
-   * The data store for the password info.
-   */
-  var data: mutable.HashMap[LoginInfo, PasswordInfo] = mutable.HashMap()
+  def remove(loginInfo: LoginInfo): Future[Unit] =
+    db.run(passwordInfoSubQuery(loginInfo).delete).map(_ => ())
 }
