+++ app/models/daos/UserDAOImpl.scala
@@ -4,15 +4,16 @@ import java.util.UUID
 
 import com.mohiva.play.silhouette.api.LoginInfo
 import models.User
-import models.daos.UserDAOImpl._
-
-import scala.collection.mutable
+import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import slick.dbio.DBIOAction
 import scala.concurrent.Future
 
 /**
- * Give access to the user object.
+ * Give access to the user object using Slick
  */
-class UserDAOImpl extends UserDAO {
+class UserDAOImpl extends UserDAO with DAOSlick {
+
+  import driver.api._
 
   /**
    * Finds a user by its login info.
@@ -20,9 +21,18 @@ class UserDAOImpl extends UserDAO {
    * @param loginInfo The login info of the user to find.
    * @return The found user or None if no user for the given login info could be found.
    */
-  def find(loginInfo: LoginInfo) = Future.successful(
-    users.find { case (id, user) => user.loginInfo == loginInfo }.map(_._2)
-  )
+  def find(loginInfo: LoginInfo) = {
+    val userQuery = for {
+      dbLoginInfo <- loginInfoQuery(loginInfo)
+      dbUserLoginInfo <- slickUserLoginInfos.filter(_.loginInfoId === dbLoginInfo.id)
+      dbUser <- slickUsers.filter(_.id === dbUserLoginInfo.userID)
+    } yield dbUser
+    db.run(userQuery.result.headOption).map { dbUserOption =>
+      dbUserOption.map { user =>
+        User(UUID.fromString(user.userID), loginInfo, user.firstName, user.lastName, user.fullName, user.email, user.avatarURL)
+      }
+    }
+  }
 
   /**
    * Finds a user by its user ID.
@@ -30,7 +40,26 @@ class UserDAOImpl extends UserDAO {
    * @param userID The ID of the user to find.
    * @return The found user or None if no user for the given ID could be found.
    */
-  def find(userID: UUID) = Future.successful(users.get(userID))
+  def find(userID: UUID) = {
+    val query = for {
+      dbUser <- slickUsers.filter(_.id === userID.toString)
+      dbUserLoginInfo <- slickUserLoginInfos.filter(_.userID === dbUser.id)
+      dbLoginInfo <- slickLoginInfos.filter(_.id === dbUserLoginInfo.loginInfoId)
+    } yield (dbUser, dbLoginInfo)
+    db.run(query.result.headOption).map { resultOption =>
+      resultOption.map {
+        case (user, loginInfo) =>
+          User(
+            UUID.fromString(user.userID),
+            LoginInfo(loginInfo.providerID, loginInfo.providerKey),
+            user.firstName,
+            user.lastName,
+            user.fullName,
+            user.email,
+            user.avatarURL)
+      }
+    }
+  }
 
   /**
    * Saves a user.
@@ -39,18 +68,28 @@ class UserDAOImpl extends UserDAO {
    * @return The saved user.
    */
   def save(user: User) = {
-    users += (user.userID -> user)
-    Future.successful(user)
+    val dbUser = DBUser(user.userID.toString, user.firstName, user.lastName, user.fullName, user.email, user.avatarURL)
+    val dbLoginInfo = DBLoginInfo(None, user.loginInfo.providerID, user.loginInfo.providerKey)
+    // We don't have the LoginInfo id so we try to get it first.
+    // If there is no LoginInfo yet for this user we retrieve the id on insertion.    
+    val loginInfoAction = {
+      val retrieveLoginInfo = slickLoginInfos.filter(
+        info => info.providerID === user.loginInfo.providerID &&
+        info.providerKey === user.loginInfo.providerKey).result.headOption
+      val insertLoginInfo = slickLoginInfos.returning(slickLoginInfos.map(_.id)).
+        into((info, id) => info.copy(id = Some(id))) += dbLoginInfo
+      for {
+        loginInfoOption <- retrieveLoginInfo
+        loginInfo <- loginInfoOption.map(DBIO.successful(_)).getOrElse(insertLoginInfo)
+      } yield loginInfo
+    }
+    // combine database actions to be run sequentially
+    val actions = (for {
+      _ <- slickUsers.insertOrUpdate(dbUser)
+      loginInfo <- loginInfoAction
+      _ <- slickUserLoginInfos += DBUserLoginInfo(dbUser.userID, loginInfo.id.get)
+    } yield ()).transactionally
+    // run actions and return user afterwards
+    db.run(actions).map(_ => user)
   }
-}
-
-/**
- * The companion object.
- */
-object UserDAOImpl {
-
-  /**
-   * The list of users.
-   */
-  val users: mutable.HashMap[UUID, User] = mutable.HashMap()
 }
