+++ app/modules/SilhouetteModule.scala
@@ -20,43 +20,28 @@ import com.mohiva.play.silhouette.impl.services._
 import com.mohiva.play.silhouette.impl.util._
 import models.User
 import models.daos._
-import models.daos.slick._
 import models.services.{ UserService, UserServiceImpl }
 import net.codingwell.scalaguice.ScalaModule
 import play.api.Play
 import play.api.Play.current
-import play.api.Logger
 import play.api.libs.concurrent.Execution.Implicits._
 import play.api.libs.openid.OpenIdClient
-import play.api.{Environment => PlayEnvironment, Configuration}
 
 /**
  * The Guice module which wires all Silhouette dependencies.
  */
-class SilhouetteModule(environment: PlayEnvironment, configuration: Configuration)
-    extends AbstractModule with ScalaModule {
+class SilhouetteModule extends AbstractModule with ScalaModule {
 
   /**
    * Configures the module.
    */
   def configure() {
     bind[UserService].to[UserServiceImpl]
-    val useSlick = configuration.getBoolean("silhouette.seed.db.useSlick").getOrElse(false)
-    if (useSlick) {
-      Logger.debug("Binding to Slick DAO implementations.")
-      bind[UserDAO].to[UserDAOSlick]
-      bind[DelegableAuthInfoDAO[PasswordInfo]].to[PasswordInfoDAOSlick]
-      bind[DelegableAuthInfoDAO[OAuth1Info]].to[OAuth1InfoDAOSlick]
-      bind[DelegableAuthInfoDAO[OAuth2Info]].to[OAuth2InfoDAOSlick]
-      bind[DelegableAuthInfoDAO[OpenIDInfo]].to[OpenIDInfoDAOSlick]
-    } else {
-      Logger.debug("Binding to In-Memory DAO implementations.")
-      bind[UserDAO].to[UserDAOImpl]
-      bind[DelegableAuthInfoDAO[PasswordInfo]].to[PasswordInfoDAO]
-      bind[DelegableAuthInfoDAO[OAuth1Info]].to[OAuth1InfoDAO]
-      bind[DelegableAuthInfoDAO[OAuth2Info]].to[OAuth2InfoDAO]
-      bind[DelegableAuthInfoDAO[OpenIDInfo]].to[OpenIDInfoDAO]
-    }
+    bind[UserDAO].to[UserDAOImpl]
+    bind[DelegableAuthInfoDAO[PasswordInfo]].to[PasswordInfoDAO]
+    bind[DelegableAuthInfoDAO[OAuth1Info]].to[OAuth1InfoDAO]
+    bind[DelegableAuthInfoDAO[OAuth2Info]].to[OAuth2InfoDAO]
+    bind[DelegableAuthInfoDAO[OpenIDInfo]].to[OpenIDInfoDAO]
     bind[CacheLayer].to[PlayCacheLayer]
     bind[HTTPLayer].toInstance(new PlayHTTPLayer)
     bind[IDGenerator].toInstance(new SecureRandomIDGenerator())
