+++ conf/application.conf
@@ -23,8 +23,6 @@ slick.dbs.default.db.url="jdbc:h2:mem:play"
 slick.dbs.default.db.user=sa
 slick.dbs.default.db.password=""
 
-silhouette.seed.db.useSlick = true
-
 # The application DI modules
 # ~~~~~
 play.modules.enabled += "modules.SilhouetteModule"
